fixing issue #25 dropped 'secret' field from the phones relation
the secret field was not used and would only confuse people. The real
secret of the sip peer is reflected in the AsteriskSip view as expected.
It is an aggregated value consiting of the first 7 characters of the md5
password hash of the associated user in corresponding room.

SCHEMA CHANGED. Database needs update:
  + removed field 'secret' from Phones relation
  + adapted asterisk-sip-view